Version 1.0 of French Translation.
This translation has been done by adapting the syntax to French language, so, the word order is sometimes different, it's "normal"that translations don't mean anything. Although I have worked for several weeks to adap to French as better as possible, it's possible that there are some mistakes made by me and I apologize for them. If that is the case, I'll correct them as soon as possible. Have a good time :)